% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.R
\name{calculate_PILSM}
\alias{calculate_PILSM}
\title{Calculate the Proportional Inter-Least Square Mean Difference (PILSM)
for One or Multiple Traits}
\usage{
calculate_PILSM(trait_values, env, covariates = NULL)
}
\arguments{
\item{trait_values}{A numeric vector containing the trait values.}

\item{env}{A vector (numeric, character, or factor) representing the environmental conditions.}

\item{covariates}{(Optional) A numeric vector or data frame for covariate values.}
}
\value{
A numeric value representing the PILSM for the trait.
}
\description{
This function calculates the Proportional Inter-Least Square Mean Difference (PILSM),
defined as the difference between the maximum and minimum least square means (LSMs)
of a trait across different environments, divided by the maximum LSM.
}
\examples{
trait_values = c(10, 12, 11, 13, 14, 15, 13, 14, 12, 13,
                 20, 22, 21, 23, 24, 25, 23, 24, 22, 23,
                 30, 32, 31, 33, 34, 35, 33, 34, 32, 33)
env = factor(rep(c("Env1", "Env2", "Env3"), each = 10))
covariates = data.frame(SoilQuality = c(rep(3, 10), rep(4, 10), rep(5, 10)))
result = calculate_PILSM(trait_values, env, covariates)
print(result)
}
