% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.R
\name{calculate_SPI}
\alias{calculate_SPI}
\title{Calculate Standardized Plasticity Index (SPI)}
\usage{
calculate_SPI(data, env_col, trait_cols, env1, env2, reference_env)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{env_col}{A string, numeric, or vector indicating the column in `data` that contains the environment labels, or an external vector specifying environment belonging.}

\item{trait_cols}{A vector of strings or numeric values indicating the columns in the dataset containing the trait values.}

\item{reference_env}{A numeric or string value representing the identifier for the reference environment in `env_col`.}

\item{trait_env1}{A numeric or string value representing the identifier for environment 1 in `env_col`.}

\item{trait_env2}{A numeric or string value representing the identifier for environment 2 in `env_col`.}
}
\value{
A named vector containing the SPI for each trait.
}
\description{
This function calculates the Standardized Plasticity Index (SPI) for multiple traits by comparing trait values
across environments relative to a reference environment.
}
\details{
The formula used is:
\deqn{SPI = \frac{Trait Value Env2 - Trait Value Env1}{Standard Deviation of Reference Trait Values}}
}
\examples{
# Example usage with synthetic data
synthetic_data = data.frame(
  Trait1_Values = c(10, 12, 15, 18),
  Trait2_Values = c(5, 6, 7, 8),
  Environment = factor(c(1, 1, 2, 2))  # Env1 = 1, Env2 = 2, Reference = 1
)
spi_values = calculate_SPI(synthetic_data, env_col = "Environment", trait_cols = c("Trait1_Values", "Trait2_Values"), trait_env1 = 1, trait_env2 = 2, reference_env = 1)
print(spi_values)

}
