% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4.R
\name{calculate_SPM}
\alias{calculate_SPM}
\title{Calculate Standardized Plasticity Metric (SPM)}
\usage{
calculate_SPM(data, env_col, trait_col, resident_env, nonresident_env)
}
\arguments{
\item{data}{A data frame containing the input data.}

\item{env_col}{A string, numeric, or vector indicating the column in `data` that contains the environment labels, or an external vector specifying environment belonging.}

\item{trait_col}{A string or numeric value indicating the column in the dataset containing the trait values.}

\item{resident_env}{A numeric or string value representing the identifier for the resident environment in `env_col`.}

\item{nonresident_env}{A numeric or string value representing the identifier for the nonresident environment in `env_col`.}
}
\value{
A numeric value representing the Standardized Plasticity Metric (SPM).
}
\description{
This function calculates the Standardized Plasticity Metric (SPM), which quantifies phenotypic plasticity
by comparing trait values between resident and nonresident environments.
}
\details{
The formula used is:
\deqn{SPM = \frac{|AinA - AinB|}{AinA} \text{ or } \frac{|BinB - BinA|}{BinB}}
}
\examples{
# Example usage with synthetic data
synthetic_data = data.frame(
  Trait_Values = c(10, 12, 15, 18),
  Environment = factor(c(1, 1, 2, 2))  # Resident = 1, Nonresident = 2
)
spm_value = calculate_SPM(synthetic_data, env_col = "Environment", trait_col = "Trait_Values", resident_env = 1, nonresident_env = 2)
print(spm_value)

}
