% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{impute}
\alias{impute}
\title{Impute Missing Values in a Data Frame or Matrix}
\usage{
impute(mt, mode = "median")
}
\arguments{
\item{mt}{A data frame or matrix containing the data to be imputed.}

\item{mode}{The method of imputation. Options are "median", "mean", "rpart", or "knn".}
}
\value{
The imputed data frame or matrix.
}
\description{
This function imputes missing values in a data frame or matrix using various methods: median, mean, rpart, or k-NN.
}
\details{
If the "rpart" method is selected, and the data frame is large, a warning is issued that the process might take a while.
If the "knn" method is selected and the VIM package is not installed, the user is prompted to install it.
}
\examples{
df = data.frame(
  Trait_1 = c(1, 2, NA, 4, 5),
  Trait_2 = c(NA, 3, 3, 4, 5),
  Trait_3 = c(10, NA, 12, 13, 14)
)
imputed_df = impute(df, mode = "median")
}
