% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1.R
\name{combine_factors}
\alias{combine_factors}
\title{Combine Internal and External Factors into a Single Dataframe and Return Levels}
\usage{
combine_factors(dataframe, factors = NULL, factors_not_in_dataframe = NULL)
}
\arguments{
\item{dataframe}{A data frame containing the data to be analyzed.}

\item{factors}{(Optional) A vector of integers or strings specifying the columns that contain the internal factors in the dataframe.
These can be the names or indices of the columns in the `dataframe`.}

\item{factors_not_in_dataframe}{(Optional) A list of vectors representing external factors that are not included in the `dataframe`.
These vectors should have the same length as the `dataframe` and correspond to the environmental conditions for each observation.}
}
\value{
A list containing:
  \item{dataframe}{The modified dataframe with a new column `Combined_Factors` representing the interaction of the provided factors.}
  \item{levels}{A list of the levels for the combined factors.}
}
\description{
This function combines internal and external factors into a single dataframe by creating
an interaction between them. It returns a modified dataframe with a new column `Combined_Factors`
that represents the interaction of these factors and a list of levels for the combined factors.
}
\examples{
# Example with internal factors
df = data.frame(
  Light = rep(c("Low", "High"), times = 6),
  Water = rep(c("Low", "High"), each = 6)
)
result = combine_factors(df, factors = c("Light", "Water"))
head(result$dataframe)
print(result$levels)

# Example with external factors
external_light = rep(c(0.4, 0.6, 0.8), each = 4)
result_external = combine_factors(df, factors_not_in_dataframe = list(external_light))
head(result_external$dataframe)
print(result_external$levels)
}
